Finish the unit tests in the POD package (#1612)
Note this is ready to merge into main, but depends on the parent PR
#1611, so this PR is targeted to that PR's feature branch. I'll deal
with the rebase after that PR is merged.

This finishes the last of the TODO(POD-P1) items by finishing the unit
tests. There was one actual bug found in the process (my POD name regex
wasn't as strict as I intended). The rest was clarification of error
messages.

I don't think it's necessary to review all the test cases in detail,
though I'd appreciate a skim and a second set of eyes on the overall
level of coverage of the test cases. Let me know if there's anything you
think is missing.

---------

Co-authored-by: Ivan Chub <ichub@users.noreply.github.com>