passport-client strict mode (#1613)
Closes
https://linear.app/0xparc-pcd/issue/0XP-202/turn-on-strict-mode-in-passport-client

Eliminates various sources of type-checking errors now that strict mode
is enabled.

Many of these are simple changes to types with no functional
implications.

Some others involve adding more explicit error states, and either
displaying user-facing errors or gracefully exiting from a process that
failed or is in a bad state (e.g. the user has edited the URL such that
query parameters are missing).

The only big change is in `pages/index.tsx`, where I've refactored the
`App` component into a functional component. This was based on
[0XP-435](https://linear.app/0xparc-pcd/issue/0XP-435/refactor-app-component-to-be-functional-rather-than-class-based)
and was necessary to solve some problems with `AppState` being
`undefined` during bootstrapping, while also using components that can
read from the `AppState`. This has been rearranged a bit, so that
`loadInitialState()` is called before mounting the `App` component, such
that everything inside `App` can rely on there being a valid state.