Fix test glob pattern quoting (#1624)
`passport-server` tests are only matching a single test file, instead of
all files in `test` and its subdirectories. This means that the majority
(227 of 231) tests for `passport-server` are not being run in CI.

Fix is the same as described here:
https://github.com/mochajs/mocha/issues/1828 - because the pattern is
unquoted, it gets expanded by the shell rather than passed through, and
the shell pattern expansion algorithm is different/inferior.