Speed up check-in requests (#1625)
Closes
https://linear.app/0xparc-pcd/issue/0XP-585/speed-up-check-in-requests
Closes
https://linear.app/0xparc-pcd/issue/0XP-610/cache-credential-verification-on-server-side

### Performance impact

Production check-in page load:

<img width="1415" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/2e448d2d-ba47-4382-8ee4-524c710b8092">

The important part here is that it takes ~4.1 seconds to get to "LCP"
(last contentful paint). This is when the pre-check has finished
loading, and we can show the user which actions they can take.

In staging-rob:
<img width="1405" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/97c3d653-d71b-4c9b-b2ca-b7111a09ba25">

The same thing takes ~1.8 seconds.

With the CORS header change in #1626, we get 1.05s (notice the lack of
the "bars" before the requests to the server, indicating that no CORS
pre-flight requests were made):
<img width="1415" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/0b273d08-cae3-402d-a45d-d2ad173ba770">
 
With server-side caching of credential verifications, we have last
contentful paint at 913ms:

<img width="1360" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/10d00f2e-b94c-4294-82ac-151f501a135b">

### What changed

#### Client-side

With this change, we now use the `CredentialManager` for both feed and
check-in/action credentials on the client. This means that when talking
to Podbox, we use the same credential in both cases. Because these
credentials are already cached for the purpose of fetching feeds, they
will almost never have to be generated as part of a check-in
(credentials do expire, so it's possible that a check-in request might
have to generate a new one, but the feed polling mechanism activates
once per minute and there's a good chance that it will have regenerated
the credential first). I have removed/replaced all instances of
"manually" created credentials, that is credentials created by
generating a payload and wrapping them in a Semaphore Signature PCD
directly. We should only use `CredentialManager` for this (or
`makeCredential` in tests).

Check-in/pre-check/action requests now include event ID, ticket ID and
action type in the request, not in the signed credential.

Devconnect check-in and offline check-in now uses the same credentials
as feeds do. This isn't a hugely consequential change, but means we
don't need to support credentials which contain only `ISSUANCE_STRING`
any more. This also means we don't need a background job to pre-generate
a special credential, since we can use the regular (cached) credentials.

#### Server-side

Credential verification is now done via the `verifyCredential()`
function in `passport-interface`, with some wrappers in
`IssuanceService` and `CredentialSubservice` for caching and some
enhanced verifications. The same mechanism is used for feeds, pre-check,
and check-in. This ensures that the full set of checks are done:

1. That the Semaphore signature is valid (`verify()` succeeds)
2. That if there is an Email PCD, it is valid (`verify()` succeeds)
3. That the Semaphore signature's identity commitment matches the Email
PCD's
4. That the timestamp on the credential has not expired

Because we previously had several different implementations of this
process, some of these checks were previously being skipped, and are now
enforced.

The return type for `verifyCredential()` is `VerifiedCredential`, which
is an interface consisting of the claims of the PCDs found in the
credential. The signature claims are always present, and claims relating
to the EmailPCD are optionally present, depending on whether the
credential included one.

When checking from `passport-server`, `verifyCredential` is wrapped in a
method on `IssuanceService`. This has had a caching mechanism in place
which has been modified to store a `VerifiedCredential`.

Podbox now has a `CredentialSubservice`, which does the same caching job
as the cache in `IssuanceService` and shares more-or-less the same
implementation (an `LRUCache`). It also has an extra method,
`verifyAndExpectZupassEmail`, which performs all of the credential
verification and also checks a) that there is an email PCD, and b) that
the email PCD is signed by Zupass. This is because it's very common in
Podbox to require this, and having one shared implementation for this
common use-case seems like a good idea. This also eliminates duplication
of verification code between pipelines.

In general I have removed all references to the concrete type
`SerializedPCD<SemaphoreSignaturePCD>` when referring to credentials,
and instead the `Credential` type from `passport-interface` is used.
Right now it aliases to `SerializedPCD<SemaphoreSignaturePCD>`, but it
makes the intent much clearer and might make it easier to adopt other
kinds of credential later. I've also updated some comments, variable
names, and method names to refer to "credentials" rather than "PCDs".

---

Overall, these changes eliminate more code than they add (the PR adds
more lines than it deletes only because it adds tests). We no longer
have lots of code on the client that looks for email PCDs, builds
payloads with timestamps, and signs these with Semaphore Signatures -
instead the same thing is achieved by calling
`credentialManager.requestCredential()`, which will use a cached
credential if one is available.

On the server-side, we also avoid deserializing two levels of PCD,
parsing JSON, verifying timestamps, and keeping track of the objects
created. Instead, we call
`credentialSubservice.verifyAndExpectZupassEmail()` or
`issuanceService.verifyCredential()`, which return a
`VerifiedCredential` that makes each PCD's claims easily accessible.