Reduce size of pcd package dependencies (#1634)
This PR reduces the bundle size impact of using most of the PCD packages
and `passport-interface`.

This is achieved by adding secondary "entry points" to these packages -
loading only part of the package's JavaScript, rather than everything in
the dependency tree of the main `src/index.ts` file. For several of the
PCD packages, that dependency tree includes heavy cryptographic
libraries. `passport-interface` also transitively depends on these, as
well as various other pieces of code that do not belong in most
third-party use-cases.

This has a drastic effect on bundle size, reducing the impact of
depending on `passport-interface` from around 2.5MB (gzipped!) to 50KB.

These savings are immediately apparent in `zupoll-client` and
`anon-message-client`, which now require 330KB and 247KB respectively on
initial load, instead of 2.53MB and 2.24MB respectively. Similar
benefits should be found in third-party applications.

For PCD packages, the `PCD` class, and types for `Args`, `Claim`,
`Proof` and so on are now under an entry point named after the PCD type.
The `PCDPackage` - which does all of the cryptographic heavy lifting and
has all of the big dependencies - is split out into a separate file.

This means that:
```typescript
// imports the whole library, including the PCDPackage
import { EdDSATicketPCDTypeName } from "@pcd/eddsa-ticket-pcd";

// imports the non-PCDPackage parts, with much fewer dependencies
import { EdDSATicketPCDTypeName } from "@pcd/eddsa-ticket-pcd/EdDSATicketPCD";
```

Examples of this are in `anon-message-client` and `zupoll-client`, which
make use of these, as well as similar changes to `passport-interface`.