passport-client re-runs background jobs when it detects that PCDCollection has changed (#1638)
I introduced this bug in
https://github.com/proofcarryingdata/zupass/pull/1625

`useCredentialManager` will return a new credential manager whenever the
PCD collection changes. However, this is not what we want in
`useBackgroundJobs`, because it will cause the hook to run again, and
the background jobs to be started multiple times! This degrades
performance substantially.

I think this has been causing several oddities and might be the reason
why we saw very bad performance on the Android devices.