GPCircuits artifacts and tests, with turborepo dependencies (#1642)
This is a redo of #1621 with changes to the turbo dependencies so that
generating test artifacts is a dependency only of `test` not `build`.
This is necessary so that the Render build/deploy doesn't need to have
`circom` installed. The way we distribute artifacts to servers for use
in prod, but it won't be based on building them with circom, so this
approach seems most appropriate, vs. installing circom on the deployment
machines.

The first of the 2 commits in this initial PR is an unmodified repeat of
#1621 so probably doesn't need further review. The 2nd commit represents
the turborepo config changes.

The core part of the approach is to make the `gen-test-artifacts` action
a dependency of `test` not `build. Only some of our packages have these
two separated out, so I modified the `@pcd/gpcircuits` package to
separate them. This is sufficient for now, but I wanted to also set
things up properly for future packages which will depend on
`@pcd/gpcircuits`, which is what the rest of this description is about.

Generalizing this a bit allows depending on the `gen-test-artifacts` of
your dependencies using `"^gen-test-artifacts". To enable this I made
the `gen-test-artifacts` action a top-level concept in our turbo and
yarn configs. This makes turbo not complain about missing actions, by
giving every package an implicit empty action. My understanding (based
on experiments) is that top-level turbo.json provides action definitions
which take effect if they are not overridden in any given package.

I added the `"^gen-test-artifacts"` dependency at the top level, and
also in the template for new packages. Any package which has that
dependency should trigger artifact generation if and only if they depend
on `@pcd/gpcircuits` (or more generally, on a package with its own
`gen-test-artifacts` action). This is technically slightly broader than
needed, in ways you'll only notice if you use turbo from inside a single
package (via `npx turbo run test`). We can't always be sure that the
tests of a package which depends on gpcircuits actually need the
artifacts, so this configuration assumes the answer is yes. An
alternative would be to assume no and require devs to add the
`"^gen-test-artifacts"` dependency manually only where needed. Slightly
too broad seemed safer than slight too narrow, to avoid more nasty
surprises in CI. Given that this is a cached action anyway, it should
only ever run once, so there's minimal wasted time.

I tested the dependencies locally by creating some packages `@pcd/foo`
and `@pcd/bar` which did or didn't depend on `@pcd/gpcircuits`. My next
major PR will introduce the `@pcd/gpc` package to exercise this stuff
for real, so I wouldn't be surprised if there are more tweaks to the
turbo config at that point.

---------

Co-authored-by: Ivan Chub <ivansergchub@gmail.com>