ZuAuth 2.0 (#1637)
Closes
https://linear.app/0xparc-pcd/issue/0XP-457/integrate-examples-with-yarn-workspacesturborepo
Closes
https://linear.app/0xparc-pcd/issue/0XP-711/new-zuauth-package-inside-monorepo

This is a rewrite of the original ZuAuth, which has the following aims:
- Least possible API complexity for third-party developers
- React optional, plain JS by default
- Minimal bundle size impact and dependency tree
- Good user experience (and API features necessary to support that)
- Example code showing realistic use

ZuAuth has a simplified JavaScript API, which has two parts: a
client-side component that builds a proof screen URL, and either opens a
popup window to Zupass, or redirects to Zupass, using that URL; and a
server-side component which authenticates the PCDs that come from that
proof screen.

Both the client-side and server-side components share configuration, so
the client-side is configured to make proofs only about PCDs that match
certain criteria, and the server-side also validates that the PCD meets
those criteria.

The client-side API is:
```javascript
import { zuAuthPopup } 

const result = await zuAuthPopup({
  fieldsToReveal: {
    revealAttendeeEmail: true,
    revealAttendeeName: true,
    revealEventId: true,
    revealProductId: true
  },
  watermark: "123456",
  eventMetadata: [
    {
      "pcdType": "eddsa-ticket-pcd",
      "publicKey": [
        "1d47687549cb273b6fed3493de5a954920dd0403f8c7eb67c2ff72a26fa4ab62",
        "1144ef5d44e2d8972d7ade8138629ebefb094025ebb4df00ed02e22d9b68e665"
      ],
      "eventId": "536c96f5-feb8-4938-bcac-47d4e13847c6",
      "eventName": "Lemonade staging",
      "productId": "9e39949c-b468-4c7e-a6a2-7735521f0bda",
      "productName": "GA"
    }
  ] // An array, so can contain multiple ticket specifications to match on
});
```
The ticket specifications in `eventMetadata` are transformed and fed in
to the prove screen via the request parameters, with the caveat that
providing many ticket specifications may produce an invalid URL (greater
than 2048 characters). Responsibility for avoiding this lies with the
caller. Allowing more flexible prove screen configuration for
`ZKEdDSAEventTicketPCD` involved making some changes to its prove screen
validation function, but these are backwards-compatible with old
callers.

The result is a `PopupActionResult`, which tells the caller whether they
got a PCD, a pending PCD, or whether the popup was closed or blocked.

Assuming they got a PCD, then the result contains a stringified
serialized PCD, which can be deserialized on the client, but should
ideally be sent to a server for verification. With server-side
verification, the bundle size impact of the client-side code is fairly
minimal. From the `zuauth` example app:

```
Route (app)                              Size     First Load JS
┌ ○ /                                    2.66 kB          87 kB
├ ○ /_not-found                          881 B          85.2 kB
├ λ /api/login                           0 B                0 B
├ λ /api/logout                          0 B                0 B
└ λ /api/watermark                       0 B                0 B
+ First Load JS shared by all            84.3 kB
  ├ chunks/1dd3208c-ba2947798056e8a5.js  53.4 kB
  ├ chunks/997-c80553db4d4d48f5.js       29 kB
  └ other shared chunks (total)          1.94 kB
```
By doing verification on the server-side, we can avoid needing to load
any heavyweight crypto libraries in the browser.

The server-side has a similarly simple API:
```javascript
const pcd = await authenticate(
  pcdStr, // as received from the popup
  watermark, // should match the one used by the client
  [
    {
      "pcdType": "eddsa-ticket-pcd",
      "publicKey": [
        "1d47687549cb273b6fed3493de5a954920dd0403f8c7eb67c2ff72a26fa4ab62",
        "1144ef5d44e2d8972d7ade8138629ebefb094025ebb4df00ed02e22d9b68e665"
      ],
      "eventId": "536c96f5-feb8-4938-bcac-47d4e13847c6",
      "eventName": "Lemonade staging",
      "productId": "9e39949c-b468-4c7e-a6a2-7735521f0bda",
      "productName": "GA"
    }
  ] // should also match the configuration used on the client side
);
```

If this succeeds, it returns a `ZKEdDSAEventTicketPCD` which has been
proven to match the criteria above. If multiple ticket specifications
are provided, the ticket just needs to match one of them.

The metadata that describes the tickets can be found in Podbox:


https://github.com/proofcarryingdata/zupass/assets/20022/d8c87a20-05a1-4738-812a-6c8fde7df340

`consumer-client` also has a ZuAuth example, which allows you to paste
metadata directly in to the page.

A standalone example app has been created at `/examples/zuauth`. This
shows configuration-sharing between client and server, and server-side
generation of a watermark. I've moved the old examples folder to
`/test-packaging`, since those examples were only used for testing
packaging stuff. The new `examples` folder is integrated into the
workspace, and `yarn build` will build the examples (of which there is
currently only one, but I will add the POD example from
https://github.com/proofcarryingdata/zupass/pull/1583 this week).

In simplifying the API, I observed the pattern of creating a `/popup`
route in the client app, which then redirects to Zupass and then
receives a redirect back from Zupass before using `postMessage` to
communicate to the window's opener. This does not seem strictly
necessary, since the popup window could open directly on `zupass.org`
(or the local equivalent) and Zupass itself could post a message back
without having to redirect. The `PCDRequest` type now has a boolean
member `postMessage` that enables this, which is currently only set by
ZuAuth. This cuts out the need for an extra route and piece of
JavaScript on the client side, and speeds the authentication process up
by eliminating redirects within the popup window. Redirect-based
authentication is still possible in the main window, where no popup
window is loaded (and thus the JavaScript for doing redirects inside the
popup window is again not needed).

Todo list:
- [x] Find a permanent home for the zuauth-example app (integrate
examples directory into workspaces?)
- [x] Better names for PCD metadata objects?
- [x] Better UI for PCD metadata in Podbox client
- [x] Metadata for other pipeline types
- [x] Better name than `zupassPopupAction`??
- [x] Integrate new ZuAuth with consumer-client
- [x] Package up server-side verification?
- [x] ZuAuth tests
- [x] Upgrade example app to latest NextJS?
- [x] Make popup optional?