Issuable PODTicketPCD (#1640)
Closes https://linear.app/0xparc-pcd/issue/0XP-724/make-a-podticketpcd

I've followed the pattern from `EDdSAPCD` and `EdDSATicketPCD`, such
that we have a `PODTicketPCD` class which contains a `PODPCD` in its
proof.

I had to make a change to `PODPCDPackage`, because the types of `PODPCD`
and `PCD<PODClaim, PODProof>` are not equivalent. The latter describes a
class which has an `id`, a `type`, a `claim`, and `proof`. PODPCD has
all of those, but also has a `_pod` member, which is not accounted for
by the generic PCD class.

Because some `PCDPackage` member functions have a return type of `PCD<C,
P>`, this means that I could not do `const pcd: PODPCD = await
PODPCDPackage.prove(/* params */)`, because `PCD<PODClaim, PODProof>`
does not satisfy the constraints of the `PODPCD` class. I've solved this
by introducing a `PODPCDPackage` interface, which inherits from
`PCDPackage<PODTicketPCDClaim, PODTicketPCDProof, PODTicketPCDArgs>` but
provides narrower return types for `prove` and `deserialize`. Everything
now type-checks.

An alternative to this approach would be to *not* include the `_pod`
member in the `PODPCD` class, but instead to instantiate one as needed.
Since it's only used for verification of the PCD, it's not strictly
necessary in the most common Zupass use-cases. Where we do verify PCDs,
we often have a cache which would mitigate the need to keep the POD
object in-memory for further verifications.

---

Aside from this, everything is pretty straightforward. Podbox is able to
issue `PODTicketPCD`s from the CSV pipeline, and I'll add similar
capabilities to the Lemonade and Pretix pipelines in
https://linear.app/0xparc-pcd/issue/0XP-533/issue-both-eddsa-and-pod-tickets.
Here is a PODPCDTicket in my PCD collection:

<img width="409" alt="image"
src="https://github.com/proofcarryingdata/zupass/assets/20022/db016177-fbb6-4531-9c64-eae9aa38d53f">

I will extend the `PODTicketPCD` tests further in the subsequent issues,
and would appreciate any suggestions for things that should be tested
for.