Allow environment-specific NODE_OPTIONS so we can use more RAM in production (#1652)
Closes
https://linear.app/0xparc-pcd/issue/0XP-431/increase-max-ram-usage-for-zupass-server-in-prod

Makes NODE_OPTIONS configurable via environment-specific environment
variables, with a default set in the `package.json` that is only used if
the environment variable is not set. Here's what happens when it runs:


![image](https://github.com/proofcarryingdata/zupass/assets/20022/4455db6b-5d65-4c9c-a606-916ca99baecb)

The default max RAM in the package.json is 4096, but in the environment
it's set to 4095, and that's what gets used.