Let CSV pipelines issue tickets to anyone who requests them (#1655)
Closes
https://linear.app/0xparc-pcd/issue/0XP-764/universal-basic-ticket-issuing-pipeline

Not sure if this is the best way to do this.

I'd like to be able to set up a pipeline such that it will issue a
ticket with a given event ID and product ID to anyone who requests one.
This PR adds an option to the CSV pipeline which causes it to issue all
of its configured tickets to anyone who hits the feed. This would allow
me to set up a feed which issues one or two tickets, which could then be
used for third party ZuAuth testing/development. I would probably only
ever enable this option for that one pipeline.

This is useful because third party developers may not have any tickets,
and I want to be able to give them a constant ZuAuth configuration that
is guaranteed to work for them, with a constant public key/event
ID/product ID.