Don't re-fetch already-subscribed DEFAULT_FEED_URL listings (#1654)
Closes
https://linear.app/0xparc-pcd/issue/0XP-780/reduce-excessive-feed-listing-for-default-feeds

We were serially hitting the Podbox feed listing endpoints for each of
the `DEFAULT_FEED_URLS` as part of the sync process. This was bad
because (a) it was done serially and not in parallel and (b) is almost
certainly unnecessary if the user is already subscribed to the feed in
question.

We can avoid this by checking if the feed is already subscribed. The
inline comments explain a bit more about how this is achieved.

Should improve the speed of "syncing PCDs", "loading tickets" and the
interstitial login screen.