Shared functionality to support upcoming GPC config (#1675)
This is a subset of the experimental functionality in #1662, pulled out
for separate review and merge. I'd like to get these changes to shared
code merged quickly so @ax0 and I can deal with any conflicts. This PR
can be reviewed and merged while I work on cleanup and testing of the
GPC functionality built on top of it.

Foundational features here:
- Added test artifacts to passport-client for dev-mode only, generated
by a gen-test-artifacts action which just copies them from gpcircuits.
These aren't actively used yet, but the copy is cheap.
- Load verification keys from disk (or via network) rather than assuming
they're imported as code. This scales better to many circuits in a
family. This uses the `fastfile` library from iden3, not because it's
the cleanest option, but because it's the option already used by snarkjs
to load proving keys, so it adds no new requirements.
- Change naming scheme for circuits to put an underscore between family
name and circuit name, for easy parsing.
- Added some new utility functions in gpcircuits and pod libraries to
help the GPC config compiler.
- Other random cleanups which came up along the way.