Finalize QR code library (#1666)
Closes
https://linear.app/0xparc-pcd/issue/0XP-762/finalize-qr-code-library

This makes Strich the default scanning library, with React QR Reader as
a fallback.

If Strich cannot initialize, e.g. due to being offline, React QR Reader
will be used instead.

I've left the Scandit code in the codebase, but without being included
in the build as it's not referenced anywhere. We can remove this once
we're totally happy that we won't need to run another comparison.