Refactor ProtoPODGPC artifact and circuit generation code (#1672)
This PR refactors the ProtoPODGPC artifact and circuit generation code.
In short:
- All of the artifact and circuit generation parameters (as well as
their associated costs) have been moved to a single JSON file
(`artifacts/circuitParameters.json`) generated by a single script (run
via `yarn gen-test-artifacts` as before). Any modifications to the
circuit parameters used would be made to the `CIRCUIT_PARAMETERS`
constant. While this is still done in the form of an array of tuples,
I'm thinking once we settle on a scheme for for the parameters, this
will be a little less opaque, e.g. if we decided to exhaust all of the
`maxObjects`, `maxEntries` and `merkleMaxDepth` values currently used,
then we would have something along the lines of the following:
```
const maxObjects = [1, 3];
const maxEntries = [5, 10];
const merkleMaxDepth = [5, 8];
const CIRCUIT_PARAMETERS = maxObjects.flatMap(nObjects =>
  maxEntries.flatMap(nEntries =>
    merkleMaxDepth.flatMap(maxDepth => [[nObjects, nEntries, maxDepth]]))) as [number,number,number][];
```
This script also regenerates `circuits.json`.
- Circuit parameters are now passed around in an object
(`ProtoPODGPCParameters`) to simplify the addition/modification of
circuit parameters (cf. #1660). A convenience function for converting
this object to an array of parameters for calls to Circomkit is provided
and used where necessary.
- The constraint counts are obtained via Circomkit's calls to SnarkJS
and differ from those provided earlier, but they are directly
proportional.

Resolves 0XP-805