Fix & cleanup gpcircuit parameters (#1681)
Another PR extracted from #1662. This one fixes an issue when the new
gen-test-artifacts runs twice.

It also includes some code changes motivated by the auto-typechecks in
VSCode. Seems the TypeScript code in `scripts/` is run but never built,
so some checks are skipped.

The change to util functions is something I was working on in #1662
which impacts gpcircuits, so is included here because splitting out all
gpcircuits changes was simplest.