Upgrade zk-kit packages (#1664)
This incorporates the latest versions of all zk-kit packages, which
we're using for POD libraries. In particular this includes a change to
EdDSA key generation we've been monitoring, which was merged here:
https://github.com/privacy-scaling-explorations/zk-kit/pull/257

I expanded the podCrypto compatibility tests to run across more
different private keys to increase the chance of running across any
incompatibilities. Looks like the output of private -> public key
generation is always the same as expected.

Before merging I'll wait for confirmation from @cedoor that the key
generation fix is finalized on the zk-kit side. Last I heard there was
one more review by cryptographers underway.