Add tuple and list membership modules (#1660)
This PR adds tuple and list membership modules to the `gpcircuits`
package. This includes additions to the proto POD GPC and modifications
to the tests. All tests pass.

Some relevant notes for context:

* Tuples are represented as Poseidon hashes of arrays of entry value
hashes of some fixed length (assumed to be between 2 and 4), and the
module outputs multiple tuples as dictated by one of its parameters.
Value hashes are chosen by index, and tuples may be fed into other
tuples by appropriate indexing. An index of -1 implies a reference to 0,
the value of choice for padding. One tuple module is instantiated in the
proto POD GPC.

* List membership is checked in the usual linear way. Lists are assumed
to be padded with the first element of the list. One list membership
module is instantiated in the proto POD GPC and the value whose
membership is to be checked is referenced by index, where an index may
refer to either an entry value hash or a computed tuple hash. Again, an
index equal to -1 implies a reference to 0, though this device isn't
strictly necessary in this case.

* Some helper functions were added. A variant of `InputSelector` is
included to allow for -1 index trick mentioned above while keeping an
out-of-bounds check in place otherwise.

* Some changes were made to the circuit family parameters.

Marking this as a draft for now as the tests need to be fleshed out a
bit to cover some edge cases (e.g. the case of an empty membership
list), and I'd like to give the code (and comments) a once-over. The
design is unlikely to change much, so feel free to peruse.

---------

Co-authored-by: Andrew Twyman <artwyman@users.noreply.github.com>