Remove package-lock.json file (#1683)
We use `yarn` for package management, so npm `package-lock.json` files
cause warnings to be printed when running `yarn`. If `npm` is being used
to manage packages locally, it's also possible that you will encounter
inconsistencies between the packages that get installed by `npm` locally
and the packages installed by `yarn` in CI and as part of the production
build process. Don't use npm!