Add GPCRequirements type for circuit selection logic (#1684)
This PR adds a `GPCRequirements` type to the `gpc` package in
preparation for the tuple and list config compilers. This type is used
in place of `ProtoPODGPCCircuitParams` to specify the circuit size
requirements. As mentioned before, while inequality checks are
sufficient for most circuit parameters, the logic involved in
determining which circuit to choose when tuples are present is a bit
more involved, hence the change.

This PR also adds circuit parameters with `MAX_LISTS = MAX_LIST_ELEMENTS
= MAX_TUPLES = TUPLE_ARITY = 0` as all tests now pass. Some additional
checks have been added to the existing tests to properly truncate test
data.

Towards 0XP-793

---------

Co-authored-by: Andrew Twyman <artwyman@users.noreply.github.com>