Fix warnings from `web-worker` transitive dependency (#1709)
Closes
https://linear.app/0xparc-pcd/issue/0XP-870/fix-build-warning-from-web-worker-dependency

The warning comes from a dependency of `snarkjs` and is fixed by a minor
version upgrade, which pulls in a fixed version of the dependency.

The changes between `0.7.3` and `0.7.4` seem to be minor, mostly
relating to testing and removal of NodeJS 16 support for SnarkJS:
https://github.com/iden3/snarkjs/compare/v0.7.3...v0.7.4

The diff is a bit confusing to read as it includes build output. The
changes in the SnarkJS build seem to result from an upgrade to the
`circom_runtime` dependency, from `0.1.24` to `0.1.25`:
https://github.com/iden3/circom_runtime/releases/tag/v0.1.25

The changes are "Support negative numbers with circom2 implementation"
and "Use Uint32Array when calculating prime as is done in circom2",
which seem like they ought not to cause any problems for us (by my
limited judgement).