Show full diff between expected and actual objects in tests (#1714)
We occasionally get a flaky test in `passport-server` that involves a
comparison between two sets. Debugging this is not helped by the fact
that the printout of the set contents is truncated.

chai has a configuration option which causes it to print out the full
contents of the compared object, instead of truncating it.