Add tuple and list membership compilation to GPC package (#1707)
Resolves
https://linear.app/0xparc-pcd/issue/0XP-793/gpc-config-compiler-for-tuple-and-list

This PR adds tuple and list membership compilation to the GPC package. A
summary of the changes:
- `GPCProofEntryConfig` now has an optional `liesInLists` field which
specifies a (non-empty) list of list names that the entry ought to lie
in.
- `GPCProofConfig` now has an optional `tuples` record mapping strings
(of type `PODName`) to an appropriate configuration object. The tuple
configuration object specifies the elements that make up the tuple as
well as the names of lists which the tuple should be a member of.
- `GPCProofInputs` includes an optional `membershipLists` field with a
record mapping strings (of type `PODName`) to lists of valid values.
- Various checks to ensure that the input and configuration are
consistent and make sense, e.g. lists and tuples contain at least two
elements and the types of values in membership lists should match up
with those entries (or tuples of entries) being checked.
- Tests in the GPC package.

Marking this as a draft until I flesh out the tests a bit more (e.g.
tests for throws), but preliminary comments are welcome.

---------

Co-authored-by: Andrew Twyman <artwyman@users.noreply.github.com>