Test gpc-pcd against artifacts both locally-built and NPM-released (#1721)
#1708 made it possible for web servers/clients to use pre-built GPC
artifacts released on NPM. However the only link between the code in the
repo and the NPM packages is a manual release process. That leaves a
risk that we won't notice when making incompatible changes, and end up
with code pushed to prod which doesn't work with the NPM artifacts. I
want a test which will flag such a situation and remind us that we need
to make a new artifact package.

The simplest safety-check I could think of is a unit test which runs
against both sets of artifacts (locally built, and NPM released). I put
that test in the @pcd/gpc-pcd package, as the middle layer between the
GPC code (where we want to allow local development) and the servers. The
same test steps in `gpc-pcd.spec.ts` are now run on each set of
artifacts.

To confirm this really does catch incompatibilities, I changed the
circuits by removing the watermark, and removed all the references to it
in code. After rebuilding everything, the gpcircuits and gpc utests
passed, as did the gpc-pcd test on test artifacts, but the new test on
NPM artifacts failed as expected.

CC @ax0 FYI regarding what it'll mean if you see this test fail in
future.