Add list and tuple support to GPCPCD (#1725)
Resolves
https://linear.app/0xparc-pcd/issue/0XP-894/add-list-and-tuple-support-to-gpcpcd

This PR integrates list membership and tuple configurations into the
gpc-pcd and consumer-client packages.

Summary of changes:
- Addition of membership lists to GPC PCDs.
- Simplified JSON serialisation of membership lists. Some of the
procedures in `podUtil.ts` have been refactored to simplify this.
- Relaxation of the list membership inclusion check to disregard type
information, i.e. it only checks for inclusion of the 'raw value'. This
was necessary to allow for serialisation-related type discrepancies.
- Appropriate modifications to the consumer-client examples to include
lists and tuples.