pod image_url etc. (#1731)
https://github.com/proofcarryingdata/zupass/assets/2636237/7c6b93d5-525a-4be9-a177-4da2bf1b34e4

#### summary
A team at ETHBerlin is working on a project using POD. They want to be
let attendees add 'contact cards' to each others' zupasses. We've been
discussing internally about having an `image_url` feature for POD PCD
UI. I've gone ahead and implemented an MVP here that should unblock
them. If we don't like how I've implemented it here, we can always
revert (this is not a one way door). I just thought it'd be nice for
them to have something a little prettier to demo.

tl;dr
- add some 'special' entry names that Zupass treats a little differently
when rendering PODs as cards
- modify the consumer-client POD example section to have an extra
example - a POD that is rendered using this new 'collectable' format.
- small unrelated change to shrink the 'change password' modal contents
- it was full width which looked a bit ridiculous